ModSettings: more changes
* Use comments as tooltips
* Use a fixed section to hide the most advanced settings from the GUI
* Use a version number to decide if preserve settings from the previous
version of the mod or recreate them
* code organization
