+++ Assets/Game/Addons/ModSupport/ModLoaderInterfaceWindow.cs
@@ -339,7 +339,7 @@ public class ModLoaderInterfaceWindow : DaggerfallPopupWindow
             modList.SelectedTextColor = Color.red;
 
         // Update buttons
-        if (ModSettingsWindow.HasSettings(ModManager.Instance.GetMod(modList.SelectedIndex)))
+        if (ModSettingsReader.HasSettings(ModManager.Instance.GetMod(modList.SelectedIndex)))
         {
             modSettingsButton.Enabled = true;
             showModDescriptionButton.Position = new Vector2(5, 83);
