+++ Assets/Game/Addons/ModSupport/ModSettings/ModSettingsReader.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: f5204f295e8ec8148affd274a75c1863
+timeCreated: 1491235491
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
